# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ExplainRequest {
  /** The type of data (e.g. location, contacts, email) */
  data_type: string;
  /** The purpose of data collection */
  purpose: string;
}

/**
 * Risk level of sharing this data
 */
export type ExplainResponseRisk =
  (typeof ExplainResponseRisk)[keyof typeof ExplainResponseRisk];

export const ExplainResponseRisk = {
  Low: "Low",
  Moderate: "Moderate",
  High: "High",
} as const;

export interface ExplainResponse {
  /** Plain-language explanation of data usage */
  explanation: string;
  /** Risk level of sharing this data */
  risk: ExplainResponseRisk;
  data_type: string;
  purpose: string;
}

export type DecisionRequestStatus =
  (typeof DecisionRequestStatus)[keyof typeof DecisionRequestStatus];

export const DecisionRequestStatus = {
  accepted: "accepted",
  rejected: "rejected",
} as const;

export interface DecisionRequest {
  status: DecisionRequestStatus;
}

export type DecisionResponseStatus =
  (typeof DecisionResponseStatus)[keyof typeof DecisionResponseStatus];

export const DecisionResponseStatus = {
  accepted: "accepted",
  rejected: "rejected",
} as const;

export interface DecisionResponse {
  message: string;
  status: DecisionResponseStatus;
}

export type StatusResponseStatus =
  (typeof StatusResponseStatus)[keyof typeof StatusResponseStatus];

export const StatusResponseStatus = {
  accepted: "accepted",
  rejected: "rejected",
  pending: "pending",
  revoked: "revoked",
} as const;

export interface StatusResponse {
  status: StatusResponseStatus;
  data_type?: string | null;
  purpose?: string | null;
}

export type RevokeResponseStatus =
  (typeof RevokeResponseStatus)[keyof typeof RevokeResponseStatus];

export const RevokeResponseStatus = {
  revoked: "revoked",
} as const;

export interface RevokeResponse {
  message: string;
  status: RevokeResponseStatus;
}