# lib/api-zod/src/generated/types/decisionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DecisionRequestStatus } from "./decisionRequestStatus";

export interface DecisionRequest {
  status: DecisionRequestStatus;
}