# lib/api-zod/src/generated/types/decisionResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DecisionResponseStatus } from "./decisionResponseStatus";

export interface DecisionResponse {
  message: string;
  status: DecisionResponseStatus;
}