# lib/api-zod/src/generated/types/explainRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExplainRequest {
  /** The type of data (e.g. location, contacts, email) */
  data_type: string;
  /** The purpose of data collection */
  purpose: string;
}