# lib/api-zod/src/generated/types/explainResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExplainResponseRisk } from "./explainResponseRisk";

export interface ExplainResponse {
  /** Plain-language explanation of data usage */
  explanation: string;
  /** Risk level of sharing this data */
  risk: ExplainResponseRisk;
  data_type: string;
  purpose: string;
}