# lib/api-zod/src/generated/types/revokeResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RevokeResponseStatus } from "./revokeResponseStatus";

export interface RevokeResponse {
  message: string;
  status: RevokeResponseStatus;
}