# lib/api-zod/src/generated/types/statusResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StatusResponseStatus } from "./statusResponseStatus";

export interface StatusResponse {
  status: StatusResponseStatus;
  data_type?: string | null;
  purpose?: string | null;
}